// src/app/models/usuario.ts
export class usuario {
    
    id: number =0
    idCuenta:number =0;
    nombre: string ="";
    genero: string="";
    telefono: string="";
    correo: string ="";
    direccion: string=""; 
    tipoUsuario: number =0;
    contrasena: string="";
    estado:number =0;
    date:string ="";
    constructor() {
    }
}